refactor(core): modularize runtime: extract debug/mood helpers, add profile builder, return boolean from pushIntentDebug, rename MAX_INTENT_HISTORY to ENTRANCE_INTENT_CO
